+++ models/Strategy.py
@@ -71,14 +71,14 @@ class Strategy:
             return False
 
         ## if last_action was set to "WAIT" due to an API problem, do not buy
-        if ( self.state.last_action == "WAIT"):
-            return False
-            log_text = (f"{str(now)} | {self.app.getMarket()} | {self.app.printGranularity()} | last_action is WAIT, do not buy yet")
+        if self.state.last_action == "WAIT":
+            log_text = f"{str(now)} | {self.app.getMarket()} | {self.app.printGranularity()} | last_action is WAIT, do not buy yet"
             Logger.warning(log_text)
+            return False
 
         # if EMA, MACD are disabled, do not buy
         if self.app.disableBuyEMA() and self.app.disableBuyMACD():
-            log_text = (f"{str(now)} | {self.app.getMarket()} | {self.app.printGranularity()} | EMA, MACD indicators are disabled")
+            log_text = f"{str(now)} | {self.app.getMarket()} | {self.app.printGranularity()} | EMA, MACD indicators are disabled"
             Logger.warning(log_text)
 
             return False
@@ -200,10 +200,10 @@ class Strategy:
                 and ((self.app.nosellmaxpcnt is not None)
                 and (margin <= self.app.nosellmaxpcnt)
             )):
-                if not app.isSimulation() or (app.isSimulation() and not app.simResultOnly()):
-                    log_text = "! Ignore Sell Signal (Within No-Sell Bounds)"
-                    Logger.warning(log_text)
-                return False
+            if not app.isSimulation() or (app.isSimulation() and not app.simResultOnly()):
+                log_text = "! Ignore Sell Signal (Within No-Sell Bounds)"
+                Logger.warning(log_text)
+            return False
 
         if debug:
             Logger.warning("\n*** isSellTrigger ***\n")